Move MEF component into package
