+++ src/ColumnGuidePackage/ColumnGuidePackage.csproj
@@ -26,6 +26,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.1.0" />
+    <PackageReference Include="Microsoft.VisualStudio.CodingConventions" Version="1.1.20180503.2" />
     <PackageReference Include="Microsoft.VSSDK.BuildTools" Version="16.4.1057" PrivateAssets="all" />
     <PackageReference Include="VSSDK.ComponentModelHost.11" Version="11.0.4" />
     <PackageReference Include="VSSDK.DTE.8" Version="8.0.4" />
@@ -35,10 +36,6 @@
     <PackageReference Include="VSSDK.Shell.Interop.11" Version="11.0.4" />
     <PackageReference Include="VSSDK.Text.11" Version="11.0.4" />
     <PackageReference Include="VSSDK.TextManager.Interop.10" Version="10.0.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ColumnGuide\ColumnGuide.csproj" />
   </ItemGroup>
 
   <ItemGroup>
