+++ src/ColumnGuidePackage/GuidelineBrush.cs
