+++ src/ColumnGuidePackage/GuidelineColorDefinition.cs
