+++ src/ColumnGuidePackage/HostServices.cs
