+++ src/ColumnGuidePackage/ITextBufferExtensions.cs
