+++ src/ColumnGuidePackage/ITextEditorGuidesSettings.cs
