+++ src/ColumnGuidePackage/ITextEditorGuidesSettingsChanger.cs
