+++ src/ColumnGuidePackage/IWpfTextViewExtensions.cs
