+++ src/ColumnGuidePackage/LineStyle.cs
