+++ src/ColumnGuidePackage/Resources.resx
@@ -117,6 +117,10 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="AddGuidelineParameterOutOfRange" xml:space="preserve">
+    <value>The parameter must be between 0 and 10,000</value>
+    <comment>Error message when an invalid value is passed to AddGuideline</comment>
+  </data>
   <data name="InvalidColumn" xml:space="preserve">
     <value>Invalid column.</value>
     <comment>Error message when a bad column number is passed to add or remove column.</comment>
@@ -124,5 +128,9 @@
   <data name="MissingIVsMonitorSelectionService" xml:space="preserve">
     <value>Missing IVsMonitorSelection service.</value>
     <comment>Error message when the IVsMonitorSelection service cannot be obtained.</comment>
+  </data>
+  <data name="RemoveGuidelineParameterOutOfRange" xml:space="preserve">
+    <value>The parameter must be between 0 and 10,000</value>
+    <comment>Error message when an invalid value is passed to RemoveGuideline</comment>
   </data>
 </root>
