+++ src/ColumnGuidePackage/StrokeParameters.cs
