+++ src/ColumnGuidePackage/TextEditorGuidesSettings.cs
@@ -12,6 +12,7 @@ using System.Text;
 using static System.Globalization.CultureInfo;
 using static EditorGuidelines.Guideline;
 
+#pragma warning disable VSTHRD010 // Invoke single-threaded types on Main thread. SettingsStore is thread safe.
 namespace EditorGuidelines
 {
     [Export(typeof(ITextEditorGuidesSettings))]
@@ -298,3 +299,4 @@ namespace EditorGuidelines
         #endregion
     }
 }
+#pragma warning restore VSTHRD010 // Invoke single-threaded types on Main thread
