+++ src/Editor Guidelines.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28729.10
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ColumnGuide", "ColumnGuide\ColumnGuide.csproj", "{0C5B092D-506C-4F92-9902-205874908ACC}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ColumnGuidePackage", "ColumnGuidePackage\ColumnGuidePackage.csproj", "{B2BA96C9-1020-4D3E-9E4F-9D2525FD4DC3}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Editor Guidelines", "VSIX\Editor Guidelines.csproj", "{BD32086D-4DD4-43E4-B1DF-69DA079DA88C}"
@@ -47,10 +45,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0C5B092D-506C-4F92-9902-205874908ACC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0C5B092D-506C-4F92-9902-205874908ACC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0C5B092D-506C-4F92-9902-205874908ACC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0C5B092D-506C-4F92-9902-205874908ACC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B2BA96C9-1020-4D3E-9E4F-9D2525FD4DC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B2BA96C9-1020-4D3E-9E4F-9D2525FD4DC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2BA96C9-1020-4D3E-9E4F-9D2525FD4DC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
