+++ test/ColumnGuideTests/ColumnGuideTests.csproj
@@ -18,7 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\ColumnGuide\ColumnGuide.csproj" />
+    <ProjectReference Include="..\..\src\ColumnGuidePackage\ColumnGuidePackage.csproj" />
   </ItemGroup>
 
 </Project>
